swig: Annotate SWIG_exit() with noreturn
Android builds all host tools with -Werror=no-return, which generates a
false positive in name_regexmatch_value() if HAVE_PCRE is not present.

Fix this by annotating SWIG_exit() with noreturn so the compiler knows
that the call will not return, thus the missing return code is in fact
not a problem.

Bug: 162801433
Change-Id: I0ee1ec89eff1f3e7aaf9056dda5543734bec2980