ANDROID: Add cc_binary_host support
Support building SWIG as a host binary. Needed for to generate python
bindings to libfdt in the U-Boot project.

The swigwarn.swg file is transformed from the C error codes header. We
can't use a genrule for this file, because it must appear in the source
tree for use by code generated by SWIG, and it only needs to be
generated "once" per source import.

The swigconfig.h is usually generated by autoconf, and contains various
hard-coded paths that don't make sense for Android.

Bug: 162801433
Change-Id: Ibd59eab31af9017dab623f33c000412077620116